Transactions (Adjustments)
 For both admin and members;
On the left hand side of the top bar, it will display the total amount of all transactions, on the right side of the top bar, it will contain the total amount of the item being searched for e.g. donations for the logged admin or member.

The description column on the table will be changed to purpose of payment and will be populated with purpose of payment related data.